libcluster/FileReader: use long instead of int to pass FileReader_state*
I fail to see the point in "hiding" the fact that this is a
pointer to FileReader_state by casting it to anything else;
unfortunately, making this explicit would break the caller's
code, so the appended uses long instead of int, which at least
can be compiled on 64-bit machines.

+++ library/cluster/FileReader.h
@@ -37,13 +37,13 @@ public:
 
 /* Routines that are independent of <T>, make them not be in class */
 
-extern int FileReader_open(char *pathname);
+extern long FileReader_open(char *pathname);
 
-extern void FileReader_close(int fs_ptr);
+extern void FileReader_close(long fs_ptr);
 
-extern int FileReader_getpos(int fs_ptr);
+extern int FileReader_getpos(long fs_ptr);
 
-extern void FileReader_setpos(int fs_ptr, int pos);
+extern void FileReader_setpos(long fs_ptr, int pos);
 
 template<class T>
 static inline T FileReader_read(int fs_ptr) {
